requirements.txt
# Phase 5: Monitoring & CI/CD Dependencies
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
scipy>=1.10.0
requests>=2.31.0

# Core ML Dependencies (for integration with existing system)
scikit-learn>=1.3.0
joblib>=1.3.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing and Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0